requirements.txt
asgiref==3.5.2
channels==3.0.4
Django==4.0.4
djangorestframework==3.13.1
PyJWT==2.4.0
requests==2.27.1
